+++ CHANGELOG.md
@@ -10,6 +10,10 @@ This project does not follow SemVer, since modules are independent of each other
 ### ecs-deploy
 - Add deployment_circuit_breaker options [#230](https://github.com/dbl-works/terraform/pull/230)
 
+### elasticache
+- Enable `encryption-in-transit` by default [#231](https://github.com/dbl-works/terraform/pull/231)
+- Change default Redis version from `6` to `7` [#231](https://github.com/dbl-works/terraform/pull/231)
+
 ## [v2023.07.19]
 ### ecs-deploy
 - Allow ecs service to be deployed to private subnets. [#228](https://github.com/dbl-works/terraform/pull/228)